backend/transaction-service/src/main/java/com/bootcamp81/transactionservice/exceptions/TransactionExceptionHandler.java
package com.bootcamp81.transactionservice.exceptions;

import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;


@ControllerAdvice
public class TransactionExceptionHandler {
    @ExceptionHandler
    public ResponseEntity<TransactionErrorResponse> handleException(Exception exc)
    {
        TransactionErrorResponse error=new TransactionErrorResponse();
        error.setStatus(HttpStatus.BAD_REQUEST.value());
        error.setMessage(exc.getMessage());
        error.setTimeStamp(System.currentTimeMillis());
        return new ResponseEntity<>(error,HttpStatus.BAD_REQUEST);
    }
}
